// src/app/enums/enums_Productos/productos.enum.ts
/**
 * Contiene las url generales para conectar con las apis
 */

export enum Enums_productos {
    /**
     *  Enums especificos para los productos 
     */ 
    Productos = 'bill-producto/', 
    todos = 'obtener-todos-productos',
    Editar = 'editar-producto',
    Borrar = 'cambiar-estado-producto/',
    Crear = 'crear-producto',
    Origen = 'obtener-todos-producto-origen'
}